first value algorithm; user guide updated
