+++ site/src/main/.vuepress/config.js
@@ -423,6 +423,7 @@ var config = {
 							['Concept/Data Type','Data Type'],
 							['Concept/Encoding','Encoding'],
 							['Concept/Compression','Compression']
+							['Concept/SDT','SDT']
 						]
 					},
 					{
@@ -943,6 +944,7 @@ var config = {
 							['Concept/Data Type','数据类型'],
 							['Concept/Encoding','编码方式'],
 							['Concept/Compression','压缩方式']
+							['Concept/SDT','旋转门压缩']
 						]
 					},
 					{
