--- tsfile/src/main/java/org/apache/iotdb/tsfile/encoding/encoder/SdtEncoder.java
@@ -1,486 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.iotdb.tsfile.encoding.encoder;
-
-import java.nio.FloatBuffer;
-import java.util.ArrayList;
-import java.util.List;
-import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
-
-public class SdtEncoder {
-
-  /**
-   * the last read time and value
-   * if upperDoor >= lowerDoor meaning out of compDeviation range, will store lastReadPair
-   */
-  private long lastReadTimestamp;
-  private long lastReadLong;
-  private double lastReadDouble;
-  private int lastReadInt;
-  private float lastReadFloat;
-  
-  /**
-   * the last stored time and vlaue
-   * we compare current point against lastStoredPair
-   */
-  private long lastStoredTimestamp;
-  private long lastStoredLong;
-  private double lastStoredDouble;
-  private int lastStoredInt;
-  private float lastStoredFloat;
-  
-  /**
-   * the maximum curUpperSlope between the lastStoredPoint to the current point
-   * upperDoor can only open up
-   */
-  private double upperDoor;
-
-  /**
-   * the minimum curLowerSlope between the lastStoredPoint to the current point
-   * lowerDoor can only open downard
-   */
-  private double lowerDoor;
-
-  private List<Long> timestamps;
-  private List<Double> doubleValues;
-  private List<Long> longValues;
-  private List<Integer> intValues;
-  private List<Float> floatValues;
-
-  /**
-   * the maximum absolute difference the user set
-   * if the data's value is within compDeviation, it will be compressed and discarded
-   * after compression, it will only store out of range <time, data></> to form the trend
-   */
-  private double compDeviation;
-
-  /**
-   * the minimum time distance between two stored data points
-   * if current point time to the last stored point time distance <= compMin,
-   * current point will NOT be stored regardless of compression deviation
-   */
-  private double compMin;
-
-  /**
-   * the maximum time distance between two stored data points
-   * if current point time to the last stored point time distance >= compMax,
-   * current point will be stored regardless of compression deviation
-   */
-  private double compMax;
-
-  TSDataType dataType;
-
-  private boolean isFirstValue;
-
-  public SdtEncoder(TSDataType dataType) {
-    timestamps = new ArrayList<>();
-    upperDoor = Integer.MIN_VALUE;
-    lowerDoor = Integer.MAX_VALUE;
-    compDeviation = -1;
-    compMin = Integer.MIN_VALUE;
-    compMax = Integer.MAX_VALUE;
-    isFirstValue = true;
-    this.dataType = dataType;
-
-    switch (dataType) {
-      case INT32:
-        intValues = new ArrayList<>();
-        break;
-      case INT64:
-        longValues = new ArrayList<>();
-        break;
-      case DOUBLE:
-        doubleValues = new ArrayList<>();
-        break;
-      case FLOAT:
-        floatValues = new ArrayList<>();
-        break;
-    }
-  }
-
-  public boolean encodeFloat(long time, float value) {
-    // store the first time and value pair
-    if (isFirtValue(time, value)) {
-      return true;
-    }
-
-    // if current point to the last stored point's time distance is within compMin,
-    // will not check two doors nor store any point within the compMin time range
-    if (time - lastStoredTimestamp <= compMin) {
-      return false;
-    }
-
-    // if current point to the last stored point's time distance is larger than compMax,
-    // will reset two doors, and store current point;
-    if (time - lastStoredTimestamp >= compMax) {
-      reset(time, value);
-      return true;
-    }
-
-    double curUpperSlope = (value - lastStoredFloat - compDeviation) / (time - lastStoredTimestamp);
-    if (curUpperSlope > upperDoor) {
-      upperDoor = curUpperSlope;
-    }
-
-    double curLowerSlope = (value - lastStoredFloat + compDeviation) / (time - lastStoredTimestamp);
-    if (curLowerSlope < lowerDoor) {
-      lowerDoor = curLowerSlope;
-    }
-
-    // current point to the lastStoredPair's value exceeds compDev, will store lastReadPair and update two doors
-    if (upperDoor >= lowerDoor) {
-      lastStoredTimestamp = lastReadTimestamp;
-      lastStoredFloat = lastReadFloat;
-      upperDoor = (value - lastStoredFloat - compDeviation) / (time - lastStoredTimestamp);
-      lowerDoor = (value - lastStoredFloat + compDeviation) / (time - lastStoredTimestamp);
-      lastReadFloat = value;
-      lastReadTimestamp = time;
-      return true;
-    }
-
-    lastReadFloat = value;
-    lastReadTimestamp = time;
-    return false;
-  }
-
-  public boolean encodeLong(long time, long value) {
-    // store the first time and value pair
-    if (isFirtValue(time, value)) {
-      return true;
-    }
-
-    // if current point to the last stored point's time distance is within compMin,
-    // will not check two doors nor store any point within the compMin time range
-    if (time - lastStoredTimestamp <= compMin) {
-      return false;
-    }
-
-    // if current point to the last stored point's time distance is larger than compMax,
-    // will reset two doors, and store current point;
-    if (time - lastStoredTimestamp >= compMax) {
-      reset(time, value);
-      return true;
-    }
-
-    double curUpperSlope = (value - lastStoredLong - compDeviation) / (time - lastStoredTimestamp);
-    if (curUpperSlope > upperDoor) {
-      upperDoor = curUpperSlope;
-    }
-
-    double curLowerSlope = (value - lastStoredLong + compDeviation) / (time - lastStoredTimestamp);
-    if (curLowerSlope < lowerDoor) {
-      lowerDoor = curLowerSlope;
-    }
-
-    // current point to the lastStoredPair's value exceeds compDev, will store lastReadPair and update two doors
-    if (upperDoor >= lowerDoor) {
-      lastStoredLong = lastReadLong;
-      lastStoredTimestamp = lastReadTimestamp;
-      upperDoor = (value - lastStoredLong - compDeviation) / (time - lastStoredTimestamp);
-      lowerDoor = (value - lastStoredLong + compDeviation) / (time - lastStoredTimestamp);
-      lastReadLong = value;
-      lastReadTimestamp = time;
-      return true;
-    }
-
-    lastReadLong = value;
-    lastReadTimestamp = time;
-    return false;
-  }
-
-  public boolean encodeInt(long time, int value) {
-    // store the first time and value pair
-    if (isFirtValue(time, value)) {
-      return true;
-    }
-
-    // if current point to the last stored point's time distance is within compMin,
-    // will not check two doors nor store any point within the compMin time range
-    if (time - lastStoredTimestamp <= compMin) {
-      return false;
-    }
-
-    // if current point to the last stored point's time distance is larger than compMax,
-    // will reset two doors, and store current point;
-    if (time - lastStoredTimestamp >= compMax) {
-      reset(time, value);
-      return true;
-    }
-
-    double curUpperSlope = (value - lastStoredInt - compDeviation) / (time - lastStoredTimestamp);
-    if (curUpperSlope > upperDoor) {
-      upperDoor = curUpperSlope;
-    }
-
-    double curLowerSlope = (value - lastStoredInt + compDeviation) / (time - lastStoredTimestamp);
-    if (curLowerSlope < lowerDoor) {
-      lowerDoor = curLowerSlope;
-    }
-
-    // current point to the lastStoredPair's value exceeds compDev, will store lastReadPair and update two doors
-    if (upperDoor >= lowerDoor) {
-      lastStoredTimestamp = lastReadTimestamp;
-      lastStoredInt = lastReadInt;
-      upperDoor = (value - lastStoredInt - compDeviation) / (time - lastStoredTimestamp);
-      lowerDoor = (value - lastStoredInt + compDeviation) / (time - lastStoredTimestamp);
-      lastReadInt = value;
-      lastReadTimestamp = time;
-      return true;
-    }
-
-    lastReadInt = value;
-    lastReadTimestamp = time;
-    return false;
-  }
-
-  public boolean encodeDouble(long time, double value) {
-    // store the first time and value pair
-    if (isFirtValue(time, value)) {
-      return true;
-    }
-
-    // if current point to the last stored point's time distance is within compMin,
-    // will not check two doors nor store any point within the compMin time range
-    if (time - lastStoredTimestamp <= compMin) {
-      return false;
-    }
-
-    // if current point to the last stored point's time distance is larger than compMax,
-    // will reset two doors, and store current point;
-    if (time - lastStoredTimestamp >= compMax) {
-      reset(time, value);
-      return true;
-    }
-
-    double curUpperSlope = (value - lastStoredDouble - compDeviation) / (time - lastStoredTimestamp);
-    if (curUpperSlope > upperDoor) {
-      upperDoor = curUpperSlope;
-    }
-
-    double curLowerSlope = (value - lastStoredDouble + compDeviation) / (time - lastStoredTimestamp);
-    if (curLowerSlope < lowerDoor) {
-      lowerDoor = curLowerSlope;
-    }
-
-    // current point to the lastStoredPair's value exceeds compDev, will store lastReadPair and update two doors
-    if (upperDoor >= lowerDoor) {
-      lastStoredTimestamp = lastReadTimestamp;
-      lastStoredDouble = lastReadDouble;
-      upperDoor = (value - lastStoredDouble - compDeviation) / (time - lastStoredTimestamp);
-      lowerDoor = (value - lastStoredDouble + compDeviation) / (time - lastStoredTimestamp);
-      lastReadDouble = value;
-      lastReadTimestamp = time;
-      return true;
-    }
-
-    lastReadDouble = value;
-    lastReadTimestamp = time;
-    return false;
-  }
-
-  public void encode(long[] timestamps, double[] values) {
-    for (int i = 0; i < timestamps.length; i++) {
-      if (encodeDouble(timestamps[i], values[i])) {
-        this.timestamps.add(getTime());
-        doubleValues.add((double) getValue());
-      }
-    }
-  }
-
-  public void encode(long[] timestamps, int[] values) {
-    for (int i = 0; i < timestamps.length; i++) {
-      if (encodeInt(timestamps[i], values[i])) {
-        this.timestamps.add(getTime());
-        intValues.add((int) getValue());
-      }
-    }
-  }
-
-  public void encode(long[] timestamps, long[] values) {
-    for (int i = 0; i < timestamps.length; i++) {
-      if (encodeLong(timestamps[i], values[i])) {
-        this.timestamps.add(getTime());
-        longValues.add((long) getValue());
-      }
-    }
-  }
-
-  public void encode(long[] timestamps, float[] values) {
-    for (int i = 0; i < timestamps.length; i++) {
-      if (encodeFloat(timestamps[i], values[i])) {
-        this.timestamps.add(getTime());
-        floatValues.add((float) getValue());
-      }
-    }
-  }
-
-  private boolean isFirtValue(long time, float value) {
-    if (isFirstValue) {
-      isFirstValue = false;
-      lastReadTimestamp = time;
-      lastReadFloat = value;
-      lastStoredTimestamp = time;
-      lastStoredFloat = value;
-      return true;
-    }
-    return false;
-  }
-
-  private boolean isFirtValue(long time, long value) {
-    if (isFirstValue) {
-      isFirstValue = false;
-      lastReadTimestamp = time;
-      lastReadLong = value;
-      lastStoredTimestamp = time;
-      lastStoredLong = value;
-      return true;
-    }
-    return false;
-  }
-
-  private boolean isFirtValue(long time, int value) {
-    if (isFirstValue) {
-      isFirstValue = false;
-      lastReadTimestamp = time;
-      lastReadInt = value;
-      lastStoredTimestamp = time;
-      lastStoredInt = value;
-      return true;
-    }
-    return false;
-  }
-
-  private boolean isFirtValue(long time, double value) {
-    if (isFirstValue) {
-      isFirstValue = false;
-      lastReadTimestamp = time;
-      lastReadDouble = value;
-      lastStoredTimestamp = time;
-      lastStoredDouble = value;
-      return true;
-    }
-    return false;
-  }
-
-  private void reset() {
-    upperDoor = Integer.MIN_VALUE;
-    lowerDoor = Integer.MAX_VALUE;
-    isFirstValue = true;
-  }
-
-  /**
-   * if current point to the last stored point's time distance >= compMax, will store current point
-   * and reset two doors
-   * @param time current time
-   * @param value current value
-   */
-  private void reset(long time, long value) {
-    reset();
-    lastStoredTimestamp = time;
-    lastStoredLong = value;
-  }
-
-  private void reset(long time, double value) {
-    reset();
-    lastStoredTimestamp = time;
-    lastStoredDouble = value;
-  }
-
-  private void reset(long time, int value) {
-    reset();
-    lastStoredTimestamp = time;
-    lastStoredInt = value;
-  }
-
-  private void reset(long time, float value) {
-    reset();
-    lastStoredTimestamp = time;
-    lastStoredFloat = value;
-  }
-
-  public void setCompDeviation(double compDeviation) {
-    this.compDeviation = compDeviation;
-  }
-
-  public double getCompDeviation() {
-    return compDeviation;
-  }
-
-  public void setCompMin(double compMin) {
-    this.compMin = compMin;
-  }
-
-  public double getCompMin() {
-    return compMin;
-  }
-
-  public void setCompMax(double compMax) {
-    this.compMax = compMax;
-  }
-
-  public double getCompMax() {
-    return compMax;
-  }
-
-  public long getTime() {
-    return lastStoredTimestamp;
-  }
-
-  public Object getValue() {
-    switch (dataType) {
-      case INT32:
-        return lastStoredInt;
-      case INT64:
-        return lastStoredLong;
-      case DOUBLE:
-        return lastStoredDouble;
-      case FLOAT:
-        return lastStoredFloat;
-      default:
-        return null;
-    }
-  }
-
-  public long[] getTimestamps() {
-    return timestamps.stream().mapToLong(i -> i).toArray();
-  }
-
-  public double[] getDoubleValues() {
-    return doubleValues.stream().mapToDouble(i -> i).toArray();
-  }
-
-  public long[] getLongValues() {
-    return longValues.stream().mapToLong(i -> i).toArray();
-  }
-
-  public int[] getIntValues() {
-    return intValues.stream().mapToInt(i -> i).toArray();
-  }
-
-  public float[] getFloatValues() {
-    return floatValues.stream().collect(
-        ()-> FloatBuffer.allocate(floatValues.size()),
-        FloatBuffer::put,
-        (left, right) -> {
-          throw new UnsupportedOperationException();
-        }
-    ).array();
-  }
-}
